Fixed
Made intended type casts explicit (to be able henceforth to detect unintended type casts) and added some overflow checks.
Solved the use of a possibly uninitialised pointer in fnft__akns_scatter_matrix.c .

+++ matlab/mex_fnft_nsep.c
@@ -35,7 +35,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
     double complex * aux_spec = NULL;
     int * sheet_indices = NULL;
     int kappa;
-    int upsampling_factor = 1;
+    size_t upsampling_factor = 1;
     size_t i;
     ptrdiff_t k;
     double *re, *im;
